requirements.txt
Flask==2.3.2
Pillow==10.0.0
pandas==2.0.3
requests==2.31.0
tensorflow==2.12.0
sympy==1.12
opencv-python==4.8.0.74
matplotlib==3.7.1